# Conflicts:
#	src/main/java/com/ii/testautomation/controllers/ProjectController.java
#	src/main/java/com/ii/testautomation/service/ProjectService.java
#	src/main/java/com/ii/testautomation/service/impl/ProjectServiceImpl.java
#	src/main/resources/MessagesAndCodes.properties

+++ src/main/java/com/ii/testautomation/controllers/ProjectController.java
@@ -7,6 +7,7 @@ import com.ii.testautomation.response.common.BaseResponse;
 import com.ii.testautomation.response.common.ContentResponse;
 import com.ii.testautomation.response.common.FileResponse;
 import com.ii.testautomation.response.common.PaginatedContentResponse;
+import com.ii.testautomation.service.ModulesService;
 import com.ii.testautomation.service.ProjectService;
 import com.ii.testautomation.utils.Constants;
 import com.ii.testautomation.utils.EndpointURI;
@@ -20,6 +21,9 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
+import java.util.List;
+import org.springframework.web.multipart.MultipartFile;
+
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -31,6 +35,8 @@ import java.util.Map;
 public class ProjectController {
     @Autowired
     private ProjectService projectService;
+    @Autowired
+    private ModulesService modulesService;
     @Autowired
     private StatusCodeBundle statusCodeBundle;
 
@@ -51,7 +57,49 @@ public class ProjectController {
         return ResponseEntity.ok(new BaseResponse(RequestStatus.SUCCESS.getStatus(),
                 statusCodeBundle.getCommonSuccessCode(),
                 statusCodeBundle.getSaveProjectSuccessMessage()));
+    }
 
+
+    @PostMapping(value = EndpointURI.PROJECT_IMPORT)
+    public ResponseEntity<Object> saveProjectByImportFile(@RequestParam MultipartFile multipartFile) {
+        List<ProjectRequest> projectRequestList = projectService.importProjectFileXls(multipartFile);
+        if (projectRequestList.isEmpty()) {
+            return ResponseEntity.ok(new BaseResponse(RequestStatus.FAILURE.getStatus(),
+                    statusCodeBundle.getFailureCode(),
+                    statusCodeBundle.getProjectFileEmptyMessage()));
+        }
+        for (ProjectRequest projectRequest : projectRequestList) {
+            if (!Utils.isNotNullAndEmpty(projectRequest.getName())) {
+                return ResponseEntity.ok(new BaseResponse(RequestStatus.FAILURE.getStatus(),
+                        statusCodeBundle.getProjectFileEmptyCode(),
+                        statusCodeBundle.getProjectNameEmptyMessage()));
+            }
+            if (!Utils.isNotNullAndEmpty(projectRequest.getCode())) {
+                return ResponseEntity.ok(new BaseResponse(RequestStatus.FAILURE.getStatus(),
+                        statusCodeBundle.getProjectFileEmptyCode(),
+                        statusCodeBundle.getProjectCodeEmptyMessage()));
+            }
+            if (!Utils.isNotNullAndEmpty(projectRequest.getDescription())) {
+                return ResponseEntity.ok(new BaseResponse(RequestStatus.FAILURE.getStatus(),
+                        statusCodeBundle.getProjectFileEmptyCode(),
+                        statusCodeBundle.getProjectDescriptionEmptyMessage()));
+            }
+            if (projectService.existByProjectName(projectRequest.getName())) {
+                return ResponseEntity.ok(new BaseResponse(RequestStatus.FAILURE.getStatus(),
+                        statusCodeBundle.getProjectAlReadyExistCode(),
+                        statusCodeBundle.getProjectNameAlReadyExistMessage()));
+            }
+            if (projectService.existByProjectCode(projectRequest.getCode())) {
+                return ResponseEntity.ok(new BaseResponse(RequestStatus.FAILURE.getStatus(),
+                        statusCodeBundle.getProjectAlReadyExistCode(),
+                        statusCodeBundle.getProjectCodeAlReadyExistMessage()));
+            }
+
+        }
+        projectService.saveProjectList(projectRequestList);
+        return ResponseEntity.ok(new BaseResponse(RequestStatus.SUCCESS.getStatus(),
+                statusCodeBundle.getCommonSuccessCode(),
+                statusCodeBundle.getSaveProjectSuccessMessage()));
     }
 
     @PutMapping(value = EndpointURI.PROJECT)
@@ -113,6 +161,12 @@ public class ProjectController {
                     statusCodeBundle.getProjectNotExistCode(),
                     statusCodeBundle.getProjectNotExistsMessage()));
         }
+        if (modulesService.existsModuleByProjectId(id)) {
+            return ResponseEntity.ok(new BaseResponse(RequestStatus.FAILURE.getStatus(),
+                    statusCodeBundle.getProjectIdDependentCode(),
+                    statusCodeBundle.getProjectIdDependentMessage()));
+        }
+
         projectService.deleteProject(id);
         return ResponseEntity.ok(new BaseResponse(RequestStatus.SUCCESS.getStatus(),
                 statusCodeBundle.getCommonSuccessCode(), statusCodeBundle.getDeleteProjectSuccessMessage()
